photoAlbum

+++ app/controller/api/v1/photoController.js
@@ -1,6 +1,9 @@
 const { Photo, User } = require("../../../models")
+const jwt = require("jsonwebtoken")
 
 const handleGetPhotos = async (req, res) => {
+    const accessToken = req.cookies.accessToken
+    if (!accessToken) return res.json({ msg: "Anda tidak diijinkan, Login Terlebih dahulu!" })
     try {
         const response = await Photo.findAll({
             include: [{ model: User }]
@@ -13,6 +16,8 @@ const handleGetPhotos = async (req, res) => {
 }
 
 const handleGetOnePhoto = (req, res) => {
+    const accessToken = req.cookies.accessToken
+    if (!accessToken) return res.json({ msg: "Anda tidak diijinkan, Login Terlebih dahulu!" })
     const { id } = req.params
     Photo.findOne({
         include: [{ model: User }]
@@ -30,21 +35,33 @@ const handlePostPhoto = (req, res) => {
     Photo.create({
         title, caption, image_url: req.file.path, userId
     }).then((foto) => {
-        res.status(201).json(foto)
+        // res.status(201).json(foto)
+        res.redirect("/listphotos")
     }).catch((error) => {
         res.status(500).json(error)
     })
 }
 
-const handleDeletePhoto = (req, res) => {
+const handleDeletePhoto = async (req, res) => {
+    const accessToken = req.cookies.accessToken
+    if (!accessToken) return res.json({ msg: "Anda tidak diijinkan, Login Terlebih dahulu!" })
     const { id } = req.params
-    Photo.destroy({
-        where: { id }
-    }).then(() => {
-        res.status(200).json({ message: "Berhasil Dihapus!" })
-    }).catch((error) => {
-        res.status(500).json(error)
-    })
+    try {
+        const user = jwt.verify(accessToken, process.env.ACCESS_TOKEN_SECRET)
+        const iduser = user.userID
+        const foto = await Photo.findOne({ where: { userId: iduser } })
+        if (!foto) return res.json({ msg: 'Anda tidak boleh menghapus foto orang lain' })
+        Photo.destroy({
+            where: { id }
+        }).then(() => {
+            // res.status(200).json({ message: "Berhasil Dihapus!" })
+            res.redirect("/photouser")
+        }).catch((error) => {
+            res.status(500).json(error)
+        })
+    } catch (error) {
+        res.json(error)
+    }
 }
 
 const handleUpdatePhoto = async (req, res) => {
@@ -55,7 +72,8 @@ const handleUpdatePhoto = async (req, res) => {
     }, {
         where: { id }
     }).then(() => {
-        res.status(200).json({ message: "Berhasil Di Update" })
+        // res.status(200).json({ message: "Berhasil Di Update" })
+        res.redirect("/photouser")
     }).catch((error) => {
         res.status(500).json(error)
     })
